+++ pumpkin/src/entity/player.rs
@@ -85,9 +85,9 @@ use crate::net::{ClientPlatform, GameProfile};
 use crate::net::{DisconnectReason, PlayerConfig};
 use crate::plugin::player::player_change_world::PlayerChangeWorldEvent;
 use crate::plugin::player::player_gamemode_change::PlayerGamemodeChangeEvent;
-use crate::plugin::player::player_item_break::PlayerItemBreakEvent;
-use crate::plugin::player::player_item_damage::PlayerItemDamageEvent;
-use crate::plugin::player::player_item_mend::PlayerItemMendEvent;
+use crate::plugin::player::item_break::PlayerItemBreakEvent;
+use crate::plugin::player::item_damage::PlayerItemDamageEvent;
+use crate::plugin::player::item_mend::PlayerItemMendEvent;
 use crate::plugin::player::player_teleport::PlayerTeleportEvent;
 use crate::server::Server;
 use crate::world::World;
