+++ pumpkin/src/plugin/api/events/player/item_consume.rs
