+++ pumpkin/src/plugin/api/events/player/mod.rs
@@ -6,10 +6,10 @@ pub mod player_gamemode_change;
 pub mod player_interact_entity_event;
 pub mod player_interact_event;
 pub mod player_interact_unknown_entity_event;
-pub mod player_item_break;
-pub mod player_item_consume;
-pub mod player_item_damage;
-pub mod player_item_mend;
+pub mod item_break;
+pub mod item_consume;
+pub mod item_damage;
+pub mod item_mend;
 pub mod player_join;
 pub mod player_leave;
 pub mod player_login;
